requirements.txt
moviepy
pillow
newspaper3k
pyttsx3
requests
lxml[html_clean]
